Restrict all functions in EIP712 authenticator to EvmCryptoSpec
